Revert "refactor(session replay): move events storage logic in IDB to separate class"
